added folders and took all logic out of index.ios.js and put it in main,js
