+++ datingApp/index.ios.js
@@ -1,8 +1,4 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- */
-'use strict';
+
 import React, {
   AppRegistry,
   Component,
@@ -11,43 +7,15 @@ import React, {
   View
 } from 'react-native';
 
+import Main from './src/main.js'
+
 class datingApp extends Component {
   render() {
     return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to Raw Dog!
-        </Text>
-        <Text style={styles.instructions}>
-          "ooooo baby I like it RAW!!"
-        </Text>
-        <Text style={styles.instructions}>
-          Lets make it happen, ive included a few folders for things like components/containers and redux folders as well
-        </Text>
-      </View>
+      <Main />
     );
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: 'red',
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center'
-    // backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    fontSize: 30,
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
 
 AppRegistry.registerComponent('datingApp', () => datingApp);
